Add files via upload
Program creates a GUI; within the GUI, prompts user to either enter a Username and Password or create an "account".  Clicking the "No?" button with open a secondary GUI that prompts the user to enter a Username and Password.  On the main GUI window, there is a button that will export the updated user_list dictionary showing the Usernames(Keys) and their Passwords(Values) as a .txt file.